Add columns and ActiveModel::Validations support
By including ActiveModel::Validations in the item,
ActiveRecord-like validations can be used when working with
Dynomite.

To get this working, I needed to implement a `column` method that
allows to define what columns *should* exist in every model.
Without it, ActiveModel::Validations couldn't tell if missing data
is a developer or user mistake.